examples: Move matriz_init() from seismic.utils.py to seismic.stiffness.utils.py
