+++ examples/seismic/utils.py
@@ -4,24 +4,10 @@ from argparse import Action, ArgumentError, ArgumentParser
 from devito import error, configuration, warning
 from devito.tools import Pickable
 
-from sympy import symbols, Matrix
-
 from .source import *
 
 __all__ = ['AcquisitionGeometry', 'setup_geometry', 'seismic_args']
 
-
-def matriz_init(model):
-    def cij(i, j):
-        ii, jj = min(i, j), max(i, j)
-        if (ii == jj or (model.dim == 3 and ii < 4 and jj < 4) or
-                        (model.dim == 2 and ii < 3 and jj < 3)):
-            return symbols('C%s%s' % (ii, jj))
-        return 0
-
-    d = model.dim*2 + model.dim-2
-    Cij = [[cij(i, j) for i in range(1, d)] for j in range(1, d)]
-    return Matrix(Cij)
 
 
 def setup_geometry(model, tn, f0=0.010):
